requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
dropbox==11.36.2
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.2
Flask==2.0.1
google-api-python-client
google-auth==1.32.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud-secret-manager==2.7.2
idna==3.6
lemmy-py==0.0.2
loguru==0.7.2
openai==1.11.1
ply==3.11
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.1
pydantic-core==2.16.2
python-telegram-bot==20.7
requests==2.31.0
urllib3==2.2.0
Werkzeug==2.0.1
